requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.0
python-dotenv==1.0.1
google-generativeai==0.7.0
httpx==0.26.0
